requirements.txt
cs50
python-dateutil
Flask
Flask-Session
Flask-SQLAlchemy
psycopg2-binary
SQLAlchemy
requests
Werkzeug
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
sqlalchemy[mypy]